requirements.txt
python-telegram-bot[job-queue]>=20.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
matplotlib>=3.7.0
pillow>=10.0.0
reportlab>=4.0.0
icalendar>=5.0.0